feat: change the Navbar and the Sighbar to the HeaderBar
