+++ src/components/Header/Header.jsx
@@ -1,9 +1,6 @@
-import React, { useContext } from 'react';
+import React from 'react';
 
-import LoggedInContext from '../../contexts/LoggedInContext';
-
-import Navbar from '../Navbar/Navbar';
-import Signbar from '../Signbar/Signbar';
+import HeaderBar from '../HeaderBar/HeaderBar';
 import CustomButton from '../CustomButton/CustomButton';
 
 import headerLogoUrl from '../../images/header-logo.svg';
@@ -11,13 +8,11 @@ import headerLogoUrl from '../../images/header-logo.svg';
 import './Header.css';
 
 function Header() {
-  const { isLoggedIn } = useContext(LoggedInContext);
-
   return (
     <header className="header">
       <img className="header__logo" src={headerLogoUrl} alt="Логотип сайта" />
+      <HeaderBar />
       <CustomButton className="header__menu-button" />
-      {isLoggedIn ? <Navbar /> : <Signbar />}
     </header>
   );
 }
