--- src/components/ProfileLink/ProfileLink.css
@@ -1,30 +1,0 @@
-.profile-link {
-  padding: 5px 10px;
-
-  border: 1px solid var(--dark-color);
-  border-radius: 17px;
-}
-
-.profile-link::before {
-  display: inline-block;
-
-  width: 10px;
-  height: 10px;
-  margin-right: 5px;
-
-  content: '';
-
-  background-image: url('../../images/profile.svg');
-
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: cover;
-}
-
-.profile-link_active {
-  border: 1px solid var(--green-color);
-}
-
-.profile-link_active::before {
-  background-image: url('../../images/profile-active.svg');
-}
