--- src/components/Signbar/Signbar.css
@@ -1,3 +1,0 @@
-.signbar {
-  position: relative;
-}
